Test pre-rendering both with and without building the entry Worker
