+++ packages/vite-plugin-cloudflare/playground/prerendering/vite.config.ts
@@ -1,4 +1,3 @@
-import assert from "node:assert";
 import * as fsp from "node:fs/promises";
 import * as path from "node:path";
 import { fileURLToPath } from "node:url";
@@ -6,17 +5,11 @@ import { cloudflare } from "@cloudflare/vite-plugin";
 import { defineConfig, preview } from "vite";
 
 export default defineConfig({
-	builder: {
-		async buildApp(builder) {
-			const clientEnvironment = builder.environments.client;
-			assert(clientEnvironment, 'No "client" environment');
-			await builder.build(clientEnvironment);
-		},
-	},
 	plugins: [
 		cloudflare({
 			inspectorPort: false,
 			persistState: false,
+			viteEnvironment: { name: "ssr" },
 			experimental: {
 				prerenderWorker: {
 					config(_, { entryWorkerConfig }) {
